do not remove rar in travis env

+++ local_rar_test.sh
@@ -3,5 +3,8 @@ export LOCALRAR=1
 cp /usr/local/bin/rar ./
 php test.php
 ./integration_test.sh
-rm rar
+# do not remove the rar on travis ci
+if [[ ! -v $TRAVIS ]]; then
+  rm rar
+fi
 unset LOCALRAR
